Add files via upload
made a TON of changes, have added a way to generate testing data, implemented a BUNCH of methods - have some ideas for more. Added a wrapper method to GappedArray which is GA
Currently lookup/contains is giving me a problem ~1% of keys that I know should be in the cALLiX aren't showing up. Not sure if this is an algorithmic problem - but it's certainly interesting.
Currently can basically adjust parameters and test speed as well as accuracy. It's pretty cool.